% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_shot.R
\name{load_reactive_objects}
\alias{load_reactive_objects}
\title{Title}
\usage{
load_reactive_objects(file)
}
\arguments{
\item{file}{}
}
\description{
Title
}
\examples{
system.file(package = "shinyloadr", "inst/Rmd/test_dashboard.Rmd") \%>\% 
 load_reactive_objects()

system.file(package = "shinyloadr", "inst/Rmd/test_dashboard_no_inputs.Rmd") \%>\% 
 load_reactive_objects()

system.file(package = "shinyloadr", "inst/Rmd/test_dashboard_missing_inputs.Rmd") \%>\% 
 load_reactive_objects()

}
